+++ tests/testthat.R
@@ -1,3 +1,4 @@
 library(testthat)
+library(pfsrsdk)
 
 test_check("pfsrsdk")